+++ tests/test_app/webroot/js/test.js
@@ -1,0 +1,9 @@
+function other_alert()
+{
+    alert('Another alert');
+}
+
+$(function () {
+    var msg = 'Ehi!';
+    alert(msg);
+});